Fix Auth Bug

+++ api/bot.go
@@ -33,7 +33,7 @@ func Bot(w http.ResponseWriter, r *http.Request) {
 	bot, _ := gotgbot.NewBot(botToken, &gotgbot.BotOpts{DisableTokenCheck: true})
 
 	// Delete the webhook incase token is unauthorized.
-	if lenAllowedTokens > 0 && !slices.Contains(allowedTokens, botToken) {
+	if lenAllowedTokens > 0 && allowedTokens[0] != "" && !slices.Contains(allowedTokens, botToken) {
 		bot.DeleteWebhook(&gotgbot.DeleteWebhookOpts{}) //nolint:errcheck // It doesn't matter if it errors
 		w.WriteHeader(statusCodeSuccess)
